型をtype配下で管理するよう変更
