+++ src/types/types.ts
@@ -1,0 +1,21 @@
+export type FeaturedChefType = {
+  name: string
+  image: string
+  path: string
+}
+
+export type TrendingRecipeType = {
+  title: string
+  description: string
+  favoriteCount: number
+  image: string
+  path: string
+}
+
+export type ChefType = {
+  name: string
+  description: string
+  image: string
+  path: string
+  recipeCount: number
+}
